Added player exist check in all plus functions, fixed heist_limit count, fixed daily reward every 50

+++ LootBot_Struct.sql
@@ -2966,7 +2966,7 @@ CREATE TABLE `public_shop` (
   `time_end` timestamp NULL DEFAULT NULL,
   `notified` tinyint(1) NOT NULL DEFAULT '0',
   `public` tinyint(1) NOT NULL DEFAULT '0',
-  `massive` int(11) NOT NULL DEFAULT '1',
+  `massive` tinyint(1) NOT NULL DEFAULT '1',
   PRIMARY KEY (`id`),
   UNIQUE KEY `shop_no_duplicate` (`code`,`item_id`),
   KEY `PLAYERID` (`player_id`),
@@ -3492,4 +3492,4 @@ CREATE TABLE `travel` (
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2019-02-18  9:00:13
+-- Dump completed on 2019-02-18 15:00:11
